Add a graph based on the hash array mapped tries of the unordered-containers library
This graph should be competitive with the patricia tree in terms of
storage overhead.  Additionally, it allows arbitrary Node
types (instead of just Int).

It also stores edge labels in *sorted* lists.  This makes graph
modification slower but gives graphs a canonical form so that they can
be more efficiently compared for equality.