Add a new graph implementation: LazyHAMT
This is based on the HashMap in unordered-containers and is very
similar to the patricia tree from fgl.  It optimizes for frequent
calls to match/context, eliminating many helper calls and expensive
allocations in both.

This graph type is for graphs that are built up once and traversed
many times (via match and/or dfs).  Insertion is somewhat slower than
the patricia tree.  It should also be a bit more compact due to the
simpler internal representation (lists versus intmaps).