Revised interface
This includes an associated type that is to be used for node marking
in algorithms that require it.  Two implementations are provided
that can be used by graph implementations:

 1) A dense bitvector (based at zero)

 2) A very sparse hashtable

A third might be useful: an IntMap where each entry is a Word64.  This
would be useful for graphs that are reasonably dense, but whose IDs do
not start at zero.