Add DFS and a new graph implementation, add mkGraph to the API
The DFS algorithms are copied from fgl.  The new graph implementation
is designed to be very efficient to traverse with a DFS; marking is
done efficiently in ST and the context function is just an IntMap
lookup.  This should vastly reduce allocations during DFS.

Added mkGraph to the API so that this immutable graph can be populated
(and it is convenient for all other graphs, too).

A variant of this current immutable graph based on vector would be
desirable, especially for graphs with contiguous vertex identifiers.
Future work.